requirements.txt
streamlit
python-docx
pandas
unidecode
qrcode[pil]
Pillow

